Various changes to file structure and added json file of spells for the db.
This update mostly consists of a large JSON file holding almost all of the spells in D&D 5e. Various scripts are also being uploaded including one to grab the spells from a website. The other scripts try and fix some of my issues from the web scraper I created. Most of them fix structure issues for the JSON file, but, due to improper planning and inexperience, I had to manually edit the JSON file to add fields. JSON file was also changed to fit the 'fixture' format for easy uploading to the spells table.